requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
requests==2.31.0
PyMuPDF==1.23.8
python-docx==1.0.1
redis==5.0.1